requirements.txt
streamlit>=1.25.0
tensorflow>=2.12.0
pillow>=9.0.0
numpy>=1.23.0
streamlit-drawable-canvas>=0.9.2
